+++ src/main/java/com/blog/domain/auth/login/currentUser/CurrentUser.java
@@ -1,0 +1,11 @@
+package com.blog.domain.auth.login.currentUser;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.PARAMETER})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CurrentUser {
+}
