--- src/main/java/com/blog/global/auth/jwt/exception/HmacSignatureException.java
@@ -1,9 +1,0 @@
-package com.blog.global.auth.jwt.exception;
-
-import com.blog.global.common.exception.BusinessLogicException;
-
-public class HmacSignatureException extends BusinessLogicException {
-	public HmacSignatureException() {
-		super(400, "JWT 서명 생성 중 오류가 발생했습니다.");
-	}
-}
