--- src/main/java/com/blog/global/auth/login/dto/LoginRequestDTO.java
@@ -1,18 +1,0 @@
-package com.blog.global.auth.login.dto;
-
-public class LoginRequestDTO {
-	private String email;
-	private String password;
-
-	public LoginRequestDTO(String email, String password) {
-		this.email = email;
-		this.password = password;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-	public String getPassword(){
-		return password;
-	}
-}
